+++ NEWS.md
@@ -32,6 +32,8 @@
 
 - Added a function `selectPage()` to select a page in the table (thanks, @carlganz, #314).
 
+- Added a function `updateCaption()` to update the table caption only (thanks, @johnpauls, #155).
+
 - When all values in a numeric column are missing (`NA`), the column filter will be disabled (http://stackoverflow.com/q/31323807/559676).
 
 - Added an argument `dec.mark` to `formatCurrency()` to customize the character for the decimal point (thanks, @frajuegies, #128).